kg-mastery/api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
google-genai
python-dotenv
